refactor: split monolithic transcribe.py into focused sub-modules
Break the 1500+ line transcribe.py into 6 modules (audio, inference,
midi, result, pipeline_single, pipeline_stems) for maintainability.
Also fix madmom build deps (>=0.16, Cython build-isolation) in
Makefile/Dockerfile.dev/pyproject.toml and pin .python-version to 3.12.
