+++ pyproject.toml
@@ -55,7 +55,7 @@ youtube = [
 basic-pitch = [
     "coremltools; platform_system == 'Darwin'",
     "librosa>=0.8.0",
-    "madmom>=0.17",
+    "madmom>=0.16",
     "mir_eval>=0.6.0",
     "numpy>=1.26",
     "onnxruntime>=1.19",
@@ -66,7 +66,7 @@ basic-pitch = [
     "typing_extensions",
 ]
 madmom = [
-    "madmom>=0.17",
+    "madmom>=0.16",
 ]
 # Offline eval harness (scripts/eval_transcription.py) — scores
 # TranscribeService output against clean_midi ground truth via
